Improve testing argument handling
Before this change, a test would actually call a library function with
(kw)arguments, so we could potentially risk errors in the function
dispatching and argument handling.

We also do not have a numbers type on the command line, but the input
from the tests were of different Python types.

Now, the tests are all run with string arguments.